# openapi/v0.yaml
openapi: 3.1.0
info:
  title: Fluorite Registry API
  version: "0.1.0"
  description: >
    Registry for publishing and fetching compiled TurboWarp extensions built
    with the Fluorite Compiler. All endpoints are versioned under /v0.
servers:
  - url: https://fluorite.sdisk.us/v0
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Accepts two kinds of bearer token: session tokens (from /auth/login,
        full access, 7-day TTL) and automation tokens (from /auth/tokens,
        long-lived, restricted to scopes granted at creation — currently only
        "publish"). Scope enforcement happens server-side per-endpoint.
  parameters:
    namespaceParam:
      name: namespace
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$"
    extensionIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-zA-Z0-9](?:[a-zA-Z0-9._-]{0,63})$"
    versionParam:
      name: version
      in: path
      required: true
      schema:
        type: string
        description: A literal semver string, or "latest".
    cursorParam:
      name: cursor
      in: query
      schema:
        type: string
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated, but not permitted to perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VERSION_PENDING_REVIEW
            message:
              type: string
              example: >-
                You have another version pending review. It needs to be
                approved or rejected before you can publish again.
            field:
              type: [string, "null"]
    User:
      type: object
      required: [namespace, displayName, type, trusted]
      properties:
        namespace:
          type: string
        displayName:
          type: string
        type:
          type: string
          enum: [admin, normal]
        trusted:
          type: boolean
          description: >-
            Whether this user's publishes are auto-approved (true after their
            first-ever publish has been reviewed and approved).
...
    VersionSummary:
      type: object
      required: [version, status, createdAt, yanked, downloads]
      properties:
        version:
          type: string
        status:
          type: string
          enum: [staging, pending, published, rejected]
        createdAt:
          type: string
          format: date-time
        publishedAt:
          type: [string, "null"]
          format: date-time
        yanked:
          type: boolean
          default: false
        yankReason:
          type: [string, "null"]
        downloads:
          type: integer
          description: "Count of compiled-code fetches (Accept: application/javascript), incremented per request."
    VersionMetadata:
      allOf:
        - $ref: "#/components/schemas/VersionSummary"
        - type: object
          required: [id, name, license, description]
          properties:
            id:
              type: string
            name:
              type: string
            license:
              type: string
            description:
              type: string
    ExtensionSummary:
      type: object
      required:
        [
          namespace,
          id,
          name,
          description,
          license,
          latestVersion,
          publishedAt,
          totalDownloads,
        ]
      properties:
        namespace:
          type: string
        id:
          type: string
        name:
          type: string
        description:
          type: string
        license:
          type: string
        latestVersion:
          type: string
          description: >-
            The highest published, non-yanked version. If every published
            version is yanked, this is omitted rather than pointing at a
            yanked version.
        publishedAt:
          type: string
          format: date-time
        totalDownloads:
          type: integer
          description: Sum of downloads across all versions of this extension.
...
paths:
...
  /users/{namespace}/trust:
    patch:
      tags: [Adminspace]
      operationId: updateUserTrust
      summary: Reset whether this user's publishes are auto-approved.
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trusted]
              properties:
                trusted:
                  type: boolean
      responses:
        "200":
          description: Trust flag updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
  /extensions:
    get:
      tags: [Open]
      operationId: listExtensions
      summary: List or search published extensions.
      security: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Matches name, id, owner namespace, and manifest fields.
        - $ref: "#/components/parameters/cursorParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: A page of extensions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  extensions:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExtensionSummary"
                  nextCursor:
                    type: [string, "null"]
...
  /extensions/@{namespace}/{id}/versions/{version}:
    get:
      tags: [Open]
      operationId: getVersion
      summary: >-
        Get a version. Content negotiation determines the response:
        application/json for metadata, application/javascript for the
        compiled code. Fetching the compiled code increments its download
        counter; fetching metadata does not. Yanked versions are still
        servable (existing consumers keep working) but are flagged.
      security: []
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
        - $ref: "#/components/parameters/extensionIdParam"
        - $ref: "#/components/parameters/versionParam"
      responses:
        "200":
          description: Version metadata or compiled code, depending on Accept.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionMetadata"
            application/javascript:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Adminspace]
      operationId: updateVersion
      summary: Approve or reject a pending version. Fires version.approved or version.rejected webhooks.
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
        - $ref: "#/components/parameters/extensionIdParam"
        - $ref: "#/components/parameters/versionParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, rejected]
                reason:
                  type: string
      responses:
        "200":
          description: Status updated. A notification is sent to the owner in the same transaction.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionMetadata"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Userspace]
      operationId: deleteVersion
      summary: Remove a single version (and its blob) without touching the rest of the extension. Owner or admin.
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
        - $ref: "#/components/parameters/extensionIdParam"
        - $ref: "#/components/parameters/versionParam"
      responses:
        "204":
          description: Deleted.
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...